+++ examples/karaoke_loopback_delayed.rs
@@ -18,7 +18,7 @@ fn processing_thread(
     sample_rate: usize,
     channels: usize,
 ) {
-    let mut pipeline = VoipAec3::builder(sample_rate as i32, channels, channels)
+    let mut pipeline = VoipAec3::builder(sample_rate, channels, channels)
         .initial_delay_ms(0)
         .build()
         .expect("failed to create AEC pipeline");
